requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.6.2
pytest-bdd==8.1.0
python-dotenv==1.0.1
pytest-html==4.1.1
pytest-metadata==3.1.1
